NAND driver support added
An independent routine nand_read_skip_bad is impemented, which provides an
abstraction to the user for reading NAND memory. This routine internally calls
specific BootROM routines to achieve the task. It is done this way to save on
code footprint since the Xloader image only has 5K bytes in SRAM to run

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>