SPEAr : Keeping the Xloader image size with in 16KiB
A small page nand block is 16KiB, so the xloader image has to allocate more than
1 block in case xloader image size is larger than 16KiB. This has implications
on u-boot/linux placing offset in nand flash.

This happens when the table is placed at larger offset in eSRAM. So, this patch
moves the table just after all sections finish. This saves on the copying time
as well

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>