Added ECC support for DDR3
SPEAr13xx platforms has an option to enable ECC for DDR3 whereby it uses the
additional DDR part for storing ECC codes which can be used for error checking
and correction upto 1 bit.
This additional DDR part must be available on the board.
This patch intends to provide a support in Xloader for this feature.

Signed-off-by: Vikas Manocha <vikas.manocha@st.com>