SPEAr13xx DDR ECC: Use C3 DMA to initialize DDR for ECC
Use C3 dma to initialize the DDR memory which is required when ECC is
enabled.  The usage of C3 can be optionaly enabled for platform but it
is a must for the current (SPEAr13XX) platforms, as the TLBs allocation
has been removed for them. As a result of which the space required to
create MMU entries would not be available

Signed-off-by: Armando Visconti <armando.visconti@st.com>
Signed-off-by: Kurt Godwin <kurt.godwin@st.com>
Signed-off-by: Mirko Gardi <mirko.gardi@st.com>
Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>