SPEAr1340/DDR: Few changes in MPMC settings
MPMC settings have been changed,
   - for ecc off to disable automatic corruption
   - for ddr3 memory output drive strength (40 Ohms)
   - to disable automatic compensation block and set manually
     comzcp=comzcn="0000" (MISC compensation_ddr_cfg)

In particular the last item will be investigated on SPEAr1340. It seems
that when we enable SPEAr1340 DDR pad automatic compensation we get
often comzcp/comzcn compensation +values that seems not the best
condition for SPEAr/DDR3 system. At the moment the best setting seems to
disable the automatic comp block and set manually +comzcp=comzcn="0000",
In fact this is a similar condition we had in SPEAR1310_REVA.

Opened a bug tracker on CODEX, and designers investigation are on going.
https://codex.cro.st.com/tracker/?func=detail&aid=140773&atid=4347&group_id=1692

Signed-off-by: Mirko Gardi <mirko.gardi@st.com>