spear1340: Accomodate bootrom table version change in Xloader
Xloader uses NAND read routines from boot code to keep a reduced code
section. This information is present in boot code itself at a
predefined address in the form of a table of function pointers.

BootROM table is modified for spear1340 and is a new version 2.1. The
spear1340 xloader has to accomodate this change and initialize the nand
routines correctly based on the table version 2.1

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>