SPEAr13xx: Add DDR clock enable and reset control in low level init code.
This patch adds in the control of clock enable and reset control of the DDR.
Before accessing the DDR, appropriate voltage signals are provided to these
GPIOs to control the DDR clock enable and reset. This finds its applications
in rendering a stability to DDR when system resumes from sleep.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>
Reviewed-by: Shiraz Hashim <shiraz.hashim@st.com>