table: Remove xloader table entirely from getting generated
The xloader table id not needed anymore by the u-boot source and only lies as
an extraneous code in xloader. This patch removes table from being compiled into
the binary

It also leads to a reduction in binary size of xloader and an improvement in
load time (read booting time) for bootrom.

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>