Align xloader.lds with TEXT_BASE
TEXT_BASE defines the entry point for xloader while xloader.lds defines the
placement of other xloader code. For certain platforms, the TEST_BASE was
defined as 0xAAAA1504 but the xloader.lds places the xloader from 0xAAAA1300.
Remove this discrepancy

Also, limit the maximum size of xloader binary to 16K for all platforms

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>