SPEAr1310 (RevC): workaround: program MPMC to prevent Linux crash
SPEAr1310 RevC has a gasket implementation on each MPMC port to
configure the writes to be not buffered or reads to be split in to a
configurable number of access.

It is observed that due to this implementation logic the write then read
on same location are executed out of order i.e. read first and then
write.

This patch configures writes on port 0 and 2 (corresponding to each
cpus) to be non-bufferable and reads on same ports to split the access
to 1 word.

This would ensure proper ordering between dependent read and writes but
would impact the performance of DDR access from CPU.

Signed-off-by: Vipin Kumar <vipin.kumar@st.com>