3xx/600 DDR initialization: fix Tinit
After DDR clock becomes stable, memory controller have to wait for
Tinit before assert CKE.

MPMC has internal counter to wait for Tinit, but it starts counting
as soon as reset is removed. This can be quite before clock gets
stable.

To respect Tinit, MPMC must be reset again after clock get stable.
The code to reset MPMC is put after pad initialization, to be sure
to provide proper clock signal to DDR chip.

This patch is not required for mobileDDR on 3xx, since in this case
MPMC automatically resets the counter when START bit is set. But
used on mobileDDR does not have any drawback.

Signed-off-by: Antonio Borneo <antonio.borneo@st.com>
Signed-off-by: Guo Jun ZENG <guojun.zeng@st.com>
Acked-by: Vipin Kumar <vipin.kumar@st.com>